RealTimeWeather/requirements.txt
pyspark==3.4.3
kafka-python==2.0.2
elasticsearch==8.14.0
requests==2.31.0